requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
django-markdownx==2.0.21
django-rest-multiple-models==2.0.1
djangorestframework==3.7.3
gunicorn==19.7.1
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
